requirements.txt
streamlit
pillow
apscheduler